[Update] 패키지 구조
ViewModel과 ui측의 패키지 구조를 변경
