+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/login/LoginModule.kt
@@ -1,7 +1,7 @@
 package com.mohaeyo.mohae.di.module.login
 
 import com.mohaeyo.mohae.di.scope.FragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.LoginViewModelFactory
+import com.mohaeyo.mohae.viewmodel.login.LoginViewModelFactory
 import dagger.Module
 import dagger.Provides
