+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/MainStaticModule.kt
@@ -1,7 +1,7 @@
 package com.mohaeyo.mohae.di.module.main
 
 import com.mohaeyo.mohae.di.scope.FragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.MainViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.MainViewModelFactory
 import dagger.Module
 import dagger.Provides
