+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/feedback/FeedbackModule.kt
@@ -1,7 +1,7 @@
 package com.mohaeyo.mohae.di.module.main.feedback
 
 import com.mohaeyo.mohae.di.scope.MainFragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.FeedbackViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.feedback.FeedbackViewModelFactory
 import dagger.Module
 import dagger.Provides
