+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/group/GroupModule.kt
@@ -1,8 +1,7 @@
 package com.mohaeyo.mohae.di.module.main.group
 
-import com.mohaeyo.mohae.di.scope.FragmentScope
 import com.mohaeyo.mohae.di.scope.MainFragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.GroupViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.group.GroupViewModelFactory
 import dagger.Module
 import dagger.Provides
