+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/mypage/MyPageModule.kt
@@ -1,8 +1,7 @@
 package com.mohaeyo.mohae.di.module.main.mypage
 
-import com.mohaeyo.mohae.di.scope.FragmentScope
 import com.mohaeyo.mohae.di.scope.MainFragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.MyPageViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.mypage.MyPageViewModelFactory
 import dagger.Module
 import dagger.Provides
