+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/place/PlaceModule.kt
@@ -1,8 +1,7 @@
 package com.mohaeyo.mohae.di.module.main.place
 
-import com.mohaeyo.mohae.di.scope.FragmentScope
 import com.mohaeyo.mohae.di.scope.MainFragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.PlaceViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.place.PlaceViewModelFactory
 import dagger.Module
 import dagger.Provides
