+++ presentation/src/main/java/com/mohaeyo/mohae/di/module/main/qa/QAModule.kt
@@ -1,7 +1,7 @@
 package com.mohaeyo.mohae.di.module.main.qa
 
 import com.mohaeyo.mohae.di.scope.FragmentScope
-import com.mohaeyo.mohae.viewmodel.facotry.QAViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.qa.QAViewModelFactory
 import dagger.Module
 import dagger.Provides
