+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/feedback/FeedbackFragment.kt
@@ -1,4 +1,4 @@
-package com.mohaeyo.mohae.ui.fragment
+package com.mohaeyo.mohae.ui.fragment.main.feedback
 
 import android.os.Bundle
 import android.view.View
