+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/group/GroupFragment.kt
@@ -1,14 +1,13 @@
-package com.mohaeyo.mohae.ui.fragment
+package com.mohaeyo.mohae.ui.fragment.main.group
 
 import android.os.Bundle
 import android.view.View
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import com.mohaeyo.mohae.R
 import com.mohaeyo.mohae.base.DataBindingFragment
 import com.mohaeyo.mohae.databinding.FragmentGroupBinding
-import com.mohaeyo.mohae.viewmodel.GroupViewModel
-import com.mohaeyo.mohae.viewmodel.facotry.GroupViewModelFactory
+import com.mohaeyo.mohae.viewmodel.main.group.GroupViewModel
+import com.mohaeyo.mohae.viewmodel.main.group.GroupViewModelFactory
 import javax.inject.Inject
 
 class GroupFragment: DataBindingFragment<FragmentGroupBinding>() {
