+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/mypage/MyPageFragment.kt
@@ -1,4 +1,4 @@
-package com.mohaeyo.mohae.ui.fragment
+package com.mohaeyo.mohae.ui.fragment.main.mypage
 
 import android.os.Bundle
 import android.view.View
