+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/qa/QAFragment.kt
@@ -1,4 +1,4 @@
-package com.mohaeyo.mohae.ui.fragment
+package com.mohaeyo.mohae.ui.fragment.main.qa
 
 import android.os.Bundle
 import android.view.View
