+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.mohaeyo.mohae.viewmodel
+package com.mohaeyo.mohae.viewmodel.main
 
 import androidx.lifecycle.Lifecycle
 import com.mohaeyo.mohae.base.BaseViewModel
