+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/MainViewModelFactory.kt
@@ -1,9 +1,10 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class PlaceViewModelFactory: ViewModelProvider.Factory {
+class MainViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T
             = modelClass.getConstructor().newInstance()
+
 }
