+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/feedback/FeedbackViewModel.kt
@@ -1,10 +1,10 @@
-package com.mohaeyo.mohae.viewmodel
+package com.mohaeyo.mohae.viewmodel.main.feedback
 
 import androidx.lifecycle.Lifecycle
 import com.mohaeyo.mohae.base.BaseViewModel
 import com.mohaeyo.mohae.base.LifecycleCallback
 
-class GroupViewModel(): LifecycleCallback, BaseViewModel() {
+class FeedbackViewModel(): LifecycleCallback, BaseViewModel() {
     override fun apply(event: Lifecycle.Event) {
 
     }
