+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/feedback/FeedbackViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main.feedback
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
