+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/group/GroupViewModel.kt
@@ -1,10 +1,10 @@
-package com.mohaeyo.mohae.viewmodel
+package com.mohaeyo.mohae.viewmodel.main.group
 
 import androidx.lifecycle.Lifecycle
 import com.mohaeyo.mohae.base.BaseViewModel
 import com.mohaeyo.mohae.base.LifecycleCallback
 
-class QAViewModel(): LifecycleCallback, BaseViewModel() {
+class GroupViewModel(): LifecycleCallback, BaseViewModel() {
     override fun apply(event: Lifecycle.Event) {
 
     }
