+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/group/GroupViewModelFactory.kt
@@ -1,10 +1,9 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main.group
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class MainViewModelFactory: ViewModelProvider.Factory {
+class GroupViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T
             = modelClass.getConstructor().newInstance()
-
 }
