+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/mypage/MyPageViewModelFactory.kt
@@ -1,9 +1,9 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main.mypage
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class QAViewModelFactory: ViewModelProvider.Factory {
+class MyPageViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T
             = modelClass.getConstructor().newInstance()
 }
