+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/place/PlaceViewModelFactory.kt
@@ -1,9 +1,9 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main.place
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class GroupViewModelFactory: ViewModelProvider.Factory {
+class PlaceViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T
             = modelClass.getConstructor().newInstance()
 }
