+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/qa/QAViewModel.kt
@@ -1,10 +1,10 @@
-package com.mohaeyo.mohae.viewmodel
+package com.mohaeyo.mohae.viewmodel.main.qa
 
 import androidx.lifecycle.Lifecycle
 import com.mohaeyo.mohae.base.BaseViewModel
 import com.mohaeyo.mohae.base.LifecycleCallback
 
-class MyPageViewModel(): LifecycleCallback, BaseViewModel() {
+class QAViewModel(): LifecycleCallback, BaseViewModel() {
     override fun apply(event: Lifecycle.Event) {
 
     }
