+++ presentation/src/main/java/com/mohaeyo/mohae/viewmodel/main/qa/QAViewModelFactory.kt
@@ -1,9 +1,9 @@
-package com.mohaeyo.mohae.viewmodel.facotry
+package com.mohaeyo.mohae.viewmodel.main.qa
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class MyPageViewModelFactory: ViewModelProvider.Factory {
+class QAViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T
             = modelClass.getConstructor().newInstance()
 }
