+++ presentation/src/main/res/layout/fragment_feedback.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
                 name="vm"
-                type="com.mohaeyo.mohae.viewmodel.FeedbackViewModel"/>
+                type="com.mohaeyo.mohae.viewmodel.main.feedback.FeedbackViewModel"/>
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
