+++ presentation/src/main/res/layout/fragment_group.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
                 name="vm"
-                type="com.mohaeyo.mohae.viewmodel.GroupViewModel"/>
+                type="com.mohaeyo.mohae.viewmodel.main.group.GroupViewModel"/>
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
