+++ presentation/src/main/res/layout/fragment_place.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
                 name="vm"
-                type="com.mohaeyo.mohae.viewmodel.PlaceViewModel"/>
+                type="com.mohaeyo.mohae.viewmodel.main.place.PlaceViewModel"/>
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
