+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -6,10 +6,10 @@
 
     <fragment
             android:id="@+id/mainFragment"
-            android:name="com.mohaeyo.mohae.ui.fragment.MainFragment"
+            android:name="com.mohaeyo.mohae.ui.fragment.main.MainFragment"
             android:label="MainFragment" >
     </fragment>
-    <fragment android:id="@+id/loginFragment" android:name="com.mohaeyo.mohae.ui.fragment.LoginFragment"
+    <fragment android:id="@+id/loginFragment" android:name="com.mohaeyo.mohae.ui.fragment.login.LoginFragment"
               android:label="LoginFragment">
         <action
                 android:id="@+id/action_loginFragment_to_main_fragment"
@@ -21,7 +21,7 @@
                 app:exitAnim="@anim/fadeout"
                 app:destination="@id/signUpFragment"/>
     </fragment>
-    <fragment android:id="@+id/signUpFragment" android:name="com.mohaeyo.mohae.ui.fragment.SignUpFragment"
+    <fragment android:id="@+id/signUpFragment" android:name="com.mohaeyo.mohae.ui.fragment.signup.SignUpFragment"
               android:label="SignUpFragment"/>
